+++ src/testing/editor_ux_parity/mod.rs
@@ -14,6 +14,7 @@ mod account_usage;
 mod handler_constructor_wrappers;
 mod handler_return_usage;
 mod handler_scope;
+mod handler_wrapper_outputs;
 mod proactive_assists;
 
 struct EditorCase<'a> {